cxx-qt-gen: move attribute and path to syntax instead of parser
