+++ cxx-qt-gen/src/parser/mod.rs
@@ -3,8 +3,6 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-pub mod attribute;
-pub mod path;
 pub mod qobject;
 
 use qobject::ParsedCxxQtData;
