+++ cxx-qt-gen/src/syntax/mod.rs
@@ -3,6 +3,8 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub mod attribute;
+pub mod path;
 mod qtfile;
 mod qtitem;
