+++ cxx-qt-gen/src/syntax/path.rs
